requirements.txt
asciimatics==1.9.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Django==3.2.8
djangorestframework==3.12.4
future==0.18.2
idna==3.3
Pillow==8.4.0
pyfiglet==0.8.post1
pypiwin32==223
pytz==2021.3
pywin32==302
requests==2.26.0
sqlparse==0.4.2
startapp==0.2.0.0
todo==0.1
todolist==0.6.0
urllib3==1.26.7
wcwidth==0.2.5
